requirements.txt
boto3==1.5.32
botocore==1.10.7
pyjwt==1.6.1
